Refactor GUI and CLI files
