+++ oommftools/user_interfaces/gui/__init__.py
@@ -1,0 +1,1 @@
+from .main import *
